# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-Agent AI System",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/agent/updateAgent/{agent_id}": {
      "put": {
        "tags": [
          "agent"
        ],
        "summary": "Update Agent",
        "description": "Update an agent's settings - PROTECTED (requires authentication and ownership)\n\nAllows updating agent settings including rate limits:\n- rate_limit_per_user_per_hour: Max requests per user per hour (null = unlimited)\n- rate_limit_system_per_hour: Max total requests per hour (null = unlimited)\n- agent_status: Agent status (active/inactive)",
        "operationId": "update_agent_api_v1_agent_updateAgent__agent_id__put",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentSettingsUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/{agent_id}/avatar": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Upload Agent Avatar",
        "description": "Upload a custom avatar image for an agent - PROTECTED (requires ownership)\n\nAccepts multipart/form-data with a single `file` field.\nStores the image under the agent-avatars/ S3 prefix (publicly readable)\nand persists the permanent URL to agent_card.iconUrl in MongoDB.\n\nReturns: { \"iconUrl\": \"<permanent public URL>\" }",
        "operationId": "upload_agent_avatar_api_v1_agent__agent_id__avatar_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_agent_avatar_api_v1_agent__agent_id__avatar_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/{agent_id}/capability-issues": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Get Capability Issues",
        "description": "Get capability issues for an agent - PROTECTED (requires ownership)",
        "operationId": "get_capability_issues_api_v1_agent__agent_id__capability_issues_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status: open or resolved",
              "title": "Status"
            },
            "description": "Filter by status: open or resolved"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Max issues to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Max issues to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of issues to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of issues to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/{agent_id}/capability-issues/resolve-all": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Resolve All Capability Issues",
        "description": "Bulk resolve all open capability issues for an agent - PROTECTED",
        "operationId": "resolve_all_capability_issues_api_v1_agent__agent_id__capability_issues_resolve_all_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/capability-issues/{issue_id}/resolve": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Resolve Capability Issue",
        "description": "Resolve a single capability issue - PROTECTED (requires ownership)",
        "operationId": "resolve_capability_issue_api_v1_agent_capability_issues__issue_id__resolve_post",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Issue Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agent/getAgent/{agent_id}": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Get Agent",
        "description": "Get agent by ID - PUBLIC (authentication optional)",
        "operationId": "get_agent_api_v1_agent_getAgent__agent_id__get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentSettingsUpdateRequest": {
        "properties": {
          "rate_limit_per_user_per_hour": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate Limit Per User Per Hour"
          },
          "rate_limit_system_per_hour": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate Limit System Per Hour"
          },
          "agent_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_public": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Public"
          }
        },
        "type": "object",
        "title": "AgentSettingsUpdateRequest",
        "description": "Request model for updating agent settings (rate limits, status, visibility)."
      },
...
      "AgentStatus": {
        "type": "string",
        "enum": [
          "active",
          "inactive",
          "deleted"
        ],
        "title": "AgentStatus"
      },
...
      "Body_upload_agent_avatar_api_v1_agent__agent_id__avatar_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_agent_avatar_api_v1_agent__agent_id__avatar_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}